+++ tests/test_credentials.py
@@ -5,16 +5,16 @@ from conftest import CosmosClientMock
 from prefect import flow
 
 from prefect_azure.credentials import (
-    BlobStorageAzureCredentials,
-    CosmosDbAzureCredentials,
-    MlAzureCredentials,
+    AzureBlobStorageCredentials,
+    AzureCosmosDbCredentials,
+    AzureMlCredentials,
 )
 
 
 def test_get_service_client(blob_connection_string):
     @flow
     def test_flow():
-        client = BlobStorageAzureCredentials(
+        client = AzureBlobStorageCredentials(
             connection_string=blob_connection_string
         ).get_client()
         return client
@@ -26,7 +26,7 @@ def test_get_service_client(blob_connection_string):
 def test_get_blob_container_client(blob_connection_string):
     @flow
     def test_flow():
-        client = BlobStorageAzureCredentials(
+        client = AzureBlobStorageCredentials(
             connection_string=blob_connection_string
         ).get_container_client("container")
         return client
@@ -39,7 +39,7 @@ def test_get_blob_container_client(blob_connection_string):
 def test_get_blob_client(blob_connection_string):
     @flow
     def test_flow():
-        client = BlobStorageAzureCredentials(
+        client = AzureBlobStorageCredentials(
             connection_string=blob_connection_string
         ).get_blob_client("container", "blob")
         return client
@@ -53,7 +53,7 @@ def test_get_blob_client(blob_connection_string):
 def test_get_cosmos_client(cosmos_connection_string):
     @flow
     def test_flow():
-        client = CosmosDbAzureCredentials(
+        client = AzureCosmosDbCredentials(
             connection_string=cosmos_connection_string
         ).get_client()
         return client
@@ -65,7 +65,7 @@ def test_get_cosmos_client(cosmos_connection_string):
 def test_get_database_client(cosmos_connection_string):
     @flow
     def test_flow():
-        client = CosmosDbAzureCredentials(
+        client = AzureCosmosDbCredentials(
             connection_string=cosmos_connection_string
         ).get_database_client("database")
         return client
@@ -77,7 +77,7 @@ def test_get_database_client(cosmos_connection_string):
 def test_get_cosmos_container_client(cosmos_connection_string):
     @flow
     def test_flow():
-        client = CosmosDbAzureCredentials(
+        client = AzureCosmosDbCredentials(
             connection_string=cosmos_connection_string
         ).get_container_client("container", "database")
         return client
@@ -91,7 +91,7 @@ def test_get_workspace(monkeypatch):
 
     @flow
     def test_flow():
-        workspace = MlAzureCredentials(
+        workspace = AzureMlCredentials(
             tenant_id="tenant_id",
             service_principal_id="service_principal_id",
             service_principal_password="service_principal_password",